refactor(server): split cron.ts into lib/cron-jobs/ (PR 2/4 tech-debt)
api/cron.ts was 1053 lines with 19 inline job bodies, violating the
self-documented 800-line max from common/coding-style.md. A typo
anywhere in the file made all 19 cron handlers fail to import,
which is exactly the failure mode the 2026-06-09 broken-migration
incident hit (gotcha #25).

Strategy: extract the 6 jobs with non-trivial inline bodies (CH
queries, multi-phase delivery, batched DB writes). The remaining 13
jobs were already lib-function-call thin and stayed inline — moving
them too would have been cosmetic-only churn that doubled review burden.

Extracted (6 new files, ~700 lines moved out of cron.ts):
- lib/cron-jobs/aggregate-usage.ts            (110 lines, was 124 inline)
- lib/cron-jobs/evaluate-alerts.ts            (215 lines, was 211 inline)
- lib/cron-jobs/detect-missing-model-prices.ts (60 lines, was 71 inline)
- lib/cron-jobs/self-monitor.ts                (85 lines, was 84 inline)
- lib/cron-jobs/detect-orphan-spans.ts         (75 lines, was 60 inline)
- lib/cron-jobs/keep-warm.ts                   (40 lines, was 29 inline)

api/cron.ts is now 383 lines (down from 1053, -670 lines). Each
extracted job is a `run<Job>Job()` function returning a typed result
object. The route handler in cron.ts is a thin wrapper: auth check,
call the job, log cron run, return JSON. The legacy result shapes are
preserved verbatim so external consumers (Vercel cron + GH Actions
cron-server.yml) see no change.

Note on logCronRun: existing unit tests use vitest spies with
`toHaveBeenCalledWith(...)` which rejects an undefined 4th argument.
The 3 wrapper that report error messages branch the call shape:
ok path passes 3 args, error path passes 4. Verbose but pinned to
the test contract.

Verification:
- pnpm --filter server test: 1022/1022 (unchanged from PR A)
- pnpm --filter server typecheck: clean

Next:
  PR C — proxy/{openai,anthropic,gemini,azure}.ts shared middleware
  PR D — eval-runner.ts 1273-line file split
