+++ apps/server/src/lib/cron-jobs/keep-warm.ts
@@ -1,0 +1,41 @@
+/**
+ * /cron/keep-warm — warm Supabase + ClickHouse pools so the next user
+ * request doesn't pay cold-start latency. Runs every 5 min.
+ *
+ * Extracted from api/cron.ts. Both warmups run via Promise.allSettled —
+ * a slow dependency cannot block the other. Never throws (this fires
+ * every 5 min and a transient warmup failure is not worth alerting on).
+ *
+ * No logCronRun call: every-5-min cadence would flood cron_job_runs.
+ */
+
+import { pingClickhouse } from '../clickhouse.js'
+import { supabaseAdmin } from '../db.js'
+
+export interface KeepWarmResult {
+  ok: boolean
+  ts: string
+  durationMs: number
+  warmed: { supabase: boolean; clickhouse: boolean }
+}
+
+export async function runKeepWarmJob(): Promise<KeepWarmResult> {
+  const started = Date.now()
+  const results = await Promise.allSettled([
+    supabaseAdmin.from('organizations').select('id', { count: 'exact', head: true }).limit(1),
+    pingClickhouse(),
+  ])
+
+  const supabaseOk = results[0].status === 'fulfilled'
+  // pingClickhouse swallows its own errors and resolves to false — read
+  // the value, not the settled status.
+  const clickhouseOk =
+    results[1].status === 'fulfilled' && results[1].value === true
+
+  return {
+    ok: supabaseOk && clickhouseOk,
+    ts: new Date().toISOString(),
+    durationMs: Date.now() - started,
+    warmed: { supabase: supabaseOk, clickhouse: clickhouseOk },
+  }
+}
